\name{elucideanDistanceToNode-package}
\alias{elucideanDistanceToNode-package}
\alias{elucideanDistanceToNode}
\docType{package}
\title{
\packageTitle{elucideanDistanceToNode}
}
\description{
\packageDescription{elucideanDistanceToNode}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{elucideanDistanceToNode}

Maintainer: \packageMaintainer{elucideanDistanceToNode}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = elucideanDistanceToNode)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{elucideanDistanceToNode}}
%\section{Documentation Index}{\packageIndices{elucideanDistanceToNode}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
